requirements.txt
fastapi
uvicorn
redis
pydantic
google-genai
pillow
numpy
opencv-python-headless
SQLAlchemy
PyMuPDF
blake3
aiofiles
python-dotenv
aioboto3
psycopg2-binary
latex2mathml
python-multipart
onnxruntime
